Added Neutron info, and cleaned up vars
Added the beginning of the Neutron configuration
info. More will come on this.
Cleaned up variables for more flexibility to
define passwords. Most of the account information
will remain as is, but passwords will/should
always change. Now one can just create the
password variable in group_vars and not have to
maintain the whole user info block.

Signed-off-by: Larry Smith Jr <mrlesmithjr@gmail.com>